+++ pkg/prometheus/promcfg_test.go
@@ -4113,6 +4113,34 @@ func TestRemoteWriteConfig(t *testing.T) {
 			golden: "RemoteWriteConfig_AzureADManagedIdentity_v3.5.0.golden",
 		},
 		{
+			version: "v3.7.0",
+			remoteWrite: monitoringv1.RemoteWriteSpec{
+				URL: "http://example.com",
+				AzureAD: &monitoringv1.AzureAD{
+					Cloud: ptr.To("AzureGovernment"),
+					WorkloadIdentity: &monitoringv1.AzureWorkloadIdentity{
+						TenantID: "00000000-a12b-3cd4-e56f-000000000000",
+						ClientID: "00000000-0000-0000-0000-000000000000",
+					},
+				},
+			},
+			golden: "RemoteWriteConfig_AzureADWorkloadIdentity_v3.7.0.golden",
+		},
+		{
+			version: "v3.6.0",
+			remoteWrite: monitoringv1.RemoteWriteSpec{
+				URL: "http://example.com",
+				AzureAD: &monitoringv1.AzureAD{
+					Cloud: ptr.To("AzureGovernment"),
+					WorkloadIdentity: &monitoringv1.AzureWorkloadIdentity{
+						TenantID: "00000000-a12b-3cd4-e56f-000000000000",
+						ClientID: "00000000-0000-0000-0000-000000000000",
+					},
+				},
+			},
+			golden: "RemoteWriteConfig_AzureADWorkloadIdentity_v3.6.0.golden",
+		},
+		{
 			version: "v2.26.0",
 			remoteWrite: monitoringv1.RemoteWriteSpec{
 				URL: "http://example.com",